+++ cmd/normalize.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"errors"
 	"fmt"
 	"log"
 	"os"
@@ -17,16 +18,39 @@ var (
 		Aliases: []string{"normalise", "norm"},
 		Long: `Normalize the given JSON schema input.
 
-The normalized JSON will be printed to STDOUT.`,
-		Example:      `  schemalint path/to/schema.json > normalized.json`,
+By default, the normalized JSON will be printed to STDOUT. Use
+--output-path / -o to specify a target path. To overwrite an
+existing file, add --force.
+`,
+		Example: `  schemalint normalize schema.json > normalized.json
+
+  schemalint normalize schema.json -o normalized.json
+
+  schemalint normalize in.json -o in.json --force
+`,
 		Args:         cobra.ExactArgs(1),
 		ArgAliases:   []string{"PATH"},
 		Run:          normalizeRun,
 		SilenceUsage: true,
 	}
+
+	outputPath     string
+	forceOverwrite bool
 )
 
+func init() {
+	normalizeCmd.Flags().StringVarP(&outputPath, "output-path", "o", "", "Output file path. If not set, STDOUT will be used.")
+	normalizeCmd.Flags().BoolVar(&forceOverwrite, "force", false, "Force overwriting any existing file when using --output-path/-o.")
+}
+
 func normalizeRun(cmd *cobra.Command, args []string) {
+	if outputPath != "" {
+		_, err := os.Stat(outputPath)
+		if !forceOverwrite && !errors.Is(err, os.ErrNotExist) {
+			log.Fatal("Error: output file already exists. Apply --force to overwrite.")
+		}
+	}
+
 	path := args[0]
 	input, err := os.ReadFile(path)
 	if err != nil {
@@ -38,7 +62,17 @@ func normalizeRun(cmd *cobra.Command, args []string) {
 		log.Fatalf("Error processing file %s.\nProbably this is not valid JSON.\nDetails: %s", path, err)
 	}
 
-	// Print normalized to STDOUT.
-	// Caution: no extra white space must be added.
-	fmt.Print(string(output))
+	// Write output.
+	if outputPath != "" {
+		err := os.WriteFile(outputPath, output, 0600)
+		if err != nil {
+			log.Fatalf("Error writing to file: %s", err)
+		} else {
+			fmt.Printf("Normalized output written to %s.\n", outputPath)
+		}
+	} else {
+		// Print normalized to STDOUT.
+		// Caution: no extra white space must be added.
+		fmt.Print(string(output))
+	}
 }